connection: add printed context support
Add the possibility to add the no_printed flag when creating the sysrepo
session to avoid loading the context from the shared memory. Apply it by
default to keep the behavior as before.
The printed context is valuable to get a the sysrepo config/state
faster but some operations on the context itself do not work.
For instance printing and parsing a yang model will fail.

Link: https://github.com/sysrepo/sysrepo/blob/v4.2.10/README.md#printed-context

Signed-off-by: Jeremie Leska <jeremie.leska@6wind.com>